chore(TypeMappings): move TypeMapping class out of Hub into its own file
